refactor: 배팅 책임 분배
